ipアドレスキャッシュ機能の実装 #117

+++ bin/dipper.sh
@@ -15,6 +15,11 @@ if [ -e ${User_File} ]; then
 fi
 
 # 環境変数宣言
+export IPV4
+export IPV4_DDNS
+export IPV6
+export IPV6_DDNS
+export IP_CACHE_TIME
 export EMAIL_CHK_DDNS
 export EMAIL_CHK_ADR
 export ERR_CHK_TIME
@@ -22,41 +27,22 @@ export ERR_CHK_TIME
 # タイマーイベントを選択し、実行する
 timer_select() {
     if [ "$IPV4" = on ] || [ "$IPV6" = on ]; then
-        ./ddns_service.sh "update" &  # DDNSアップデートタイマーを開始
+        ./ddns_service.sh "update" &    # DDNSアップデートタイマーを開始
     fi
 
     if [  "$IPV4" = on ] && [ "$IPV4_DDNS" = on ]; then
-        ./ddns_service.sh "check" &  # DDNSチェックタイマーを開始
+        ./ddns_service.sh "check" &     # DDNSチェックタイマーを開始
 
     elif [ "$IPV6" = on ] && [ "$IPV6_DDNS" = on ]; then
-        ./ddns_service.sh "check" &  # DDNSチェックタイマーを開始
+        ./ddns_service.sh "check" &     # DDNSチェックタイマーを開始
     fi
 }
 
-# メール通知機能チェック処理
-mail_service() {
+dir_check() {
     local cache_dir="../cache"
-    local cache_err="${cache_dir}/err_mail"
-    local cache_ddns="${cache_dir}/ddns_mail"
-
-    if [[ -n ${EMAIL_CHK_ADR:-} ]]; then
-        if [[ -n ${ERR_CHK_TIME:-} ]]; then
-            ./err_mail_service.sh "$EMAIL_CHK_ADR" "$ERR_CHK_TIME" &
-        elif [ -f "${cache_err}" ]; then
-            rm "${cache_err}"
-        fi
-
-        if [[ -n ${EMAIL_CHK_DDNS:-} ]]; then
-            ./mail_handle.sh "ddns_mail" "IPアドレスの変更がありました <$(hostname)>" "$EMAIL_CHK_ADR" &
-        elif [ -f "${cache_ddns}" ]; then
-            rm "${cache_ddns}"
-        fi
-        # ディレクトリの中身をチェック
-        if [ -d "${cache_dir}" ] && [ -z "$(ls -A ${cache_dir})" ]; then
-            # ファイルが存在しない
-            rm -r "${cache_dir}"
-        fi
-    elif [ -d "${cache_dir}" ]; then
+    # ディレクトリの中身をチェック
+    if [ -d "${cache_dir}" ] && [ -z "$(ls -A ${cache_dir})" ]; then
+        # ファイルが存在しない
         rm -r "${cache_dir}"
     fi
 }
@@ -65,18 +51,20 @@ main() {
     local exit_code=""
 
     timer_select
-    mail_service
+    ./mail_service.sh &
+    ./cache_ip_service.sh &
+    dir_check
     # バックグラウンドプロセスを監視して通常終了以外の時、異常終了させる
     while true;do
         wait -n
         exit_code=$?
         if [ "$exit_code" = 127 ]; then
             ./err_message.sh "process" "dipper.sh" "endcode=$exit_code  プロセスが全て終了しました。"
-            ./mail_handle.sh "err_mail" "dipperでエラーを検出しました <$(hostname)>" "$EMAIL_CHK_ADR"
+            ./mail_sending.sh "err_mail" "dipperでエラーを検出しました <$(hostname)>" "$EMAIL_CHK_ADR"
             exit 0
         elif [ "$exit_code" != 0 ]; then
             ./err_message.sh "process" "dipper.sh" "endcode=$exit_code  プロセスのどれかが異常終了した為、強制終了しました。"
-            ./mail_handle.sh "err_mail" "dipperでエラーを検出しました <$(hostname)>" "$EMAIL_CHK_ADR"
+            ./mail_sending.sh "err_mail" "dipperでエラーを検出しました <$(hostname)>" "$EMAIL_CHK_ADR"
             exit 1
         fi
         sleep 10
